Refactoring of the application.
Changed the view which is now a multiadapter for a content object and a
screen. With this, I changed the application which is now seperated
between the content application object and the view of the application.
The content application obj and the view is created during running the
instance.
The quit key handler was removed. It's purpose was testing the CA
only, anyway.