Implemented a keydispatcher.
The aim to use some sort of a dispatching utility is to distinguish
which keys are pressed and assign them different functionality depending
on the widget or view. Currently a lot of this functionality is crammed
into the run function in the view, determining based on the focused
widget which receives the key stroke.