Removed the keydispatcher.
I'm now sending almost all key events to the top widget itself (except
the button). The button needs to be improved, so it's easier to
distinguis if it's clicked what to do. I tried a signal connect, but
can't stop the event loop in the view. That somehow needs a better
solution.