Finally store projects in the database.
I now use zope.container for the folder like structure, after finding
out, that I can't use or inherit from BTrees directly - stupid me. The
database is now used by the application object, which acts like a
dictionary type object. Keys and Values set to the application will be
stored in the database, acting as a root-object in the database.